Refatorar pacotes e adicionar dependências
Ajuste na estrutura

Refatorar .gitignore para adicionar diretórios antigos

Refatorar estilos do componente Navbar

Atualizar módulo de departamentos para adicionar dependência de unidades

Atualizar entidade de departamento para adicionar relacionamento com usuários

Atualizar controlador de departamentos para adicionar rota de busca sem paginação

Atualizar controlador de departamentos para adicionar rota de busca com paginação

Atualizar controlador de departamentos para adicionar rota de busca por unidade

Atualizar controlador de departamentos para adicionar rota de busca por ID

+++ backend/src/departments/departments.service.ts
@@ -1,47 +1,55 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { QueryFailedError, Repository } from 'typeorm';
 import { UpdateDepartmentDto } from './dtos/updateDepartment.dto';
 import { DepartmentEntity } from './entities/department.entity';
 import { CreateDepartmentDto } from './dtos/createDepartment.dto';
+import { UnitEntity } from '../units/entities/unit.entity'; 
 
 @Injectable()
 export class DepartmentsService {
   private readonly logger = new Logger(DepartmentsService.name);
 
   constructor(
+    @InjectRepository(UnitEntity)
+    private readonly unitRepository: Repository<UnitEntity>,
     @InjectRepository(DepartmentEntity)
     private readonly departmentRepository: Repository<DepartmentEntity>,
   ) {}
 
-  async findByUnit(unitId: string): Promise<DepartmentEntity[]> {
-    try {
-      return await this.departmentRepository.find({ where: { unit: { id: unitId } } });
-    } catch (error) {
-      this.logger.error('Erro ao buscar departamentos por unidade', error.stack);
-      throw new Error(`Erro ao buscar departamentos por unidade: ${error.message}`);
-    }
-  }
-
+  // Criação de um novo departamento
   async createDepartment(createDepartmentDto: CreateDepartmentDto): Promise<DepartmentEntity> {
     try {
-      const department = {
+      const unit = await this.unitRepository.findOne({ where: { id: createDepartmentDto.unit_id } });
+      if (!unit) {
+        throw new NotFoundException('Unidade não encontrada');
+      }
+
+      const department = this.departmentRepository.create({
         ...createDepartmentDto,
+        unit,
         created_at: new Date(),
         updated_at: new Date(),
-      };
-      return await this.departmentRepository.save(department);
+      });
+
+      this.logger.log('Criando departamento ', department);
+      await this.departmentRepository.save(department);
+      return department;
     } catch (error) {
       this.logger.error('Erro ao criar departamento', error.stack);
-      throw new Error(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      }
+      throw new BadRequestException('Erro ao criar departamento');
     }
   }
 
+  // Atualização de um departamento existente
   async updateDepartment(id: string, updateDepartmentDto: UpdateDepartmentDto): Promise<DepartmentEntity> {
     try {
       const department = await this.getDepartmentById(id);
       if (!department) {
-        throw new Error('Department not found');
+        throw new NotFoundException('Departamento não encontrado');
       }
       const updatedDepartment = {
         ...department,
@@ -51,44 +59,74 @@ export class DepartmentsService {
       return await this.departmentRepository.save(updatedDepartment);
     } catch (error) {
       this.logger.error('Erro ao atualizar departamento', error.stack);
-      throw new Error(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      }
+      throw new BadRequestException('Erro ao atualizar departamento');
     }
   }
 
-  async getAllDepartments(page: number, limit: number): Promise<{ data: DepartmentEntity[], total: number }> {
+  // Método para buscar todos os departamentos com paginação, filtro e ordenação (case-insensitive)
+  async getAllDepartmentsWithPagination(
+    page: number,
+    limit: number,
+    filter: string,
+    sortBy: string,
+    sortOrder: 'ASC' | 'DESC',
+  ): Promise<{ data: DepartmentEntity[]; total: number }> {
     try {
-      this.logger.log('Buscando todos os departamentos');
-      const [departments, total] = await this.departmentRepository.findAndCount({
-        skip: (page - 1) * limit,
-        take: limit,
-        relations: ['unit'], // Certifique-se de carregar a relação 'unit'
-      });
-      this.logger.log(`Departamentos encontrados: ${departments.length}`);
+      const skip = (page - 1) * limit;
+
+      const queryBuilder = this.departmentRepository.createQueryBuilder('department');
+
+      if (filter) {
+        queryBuilder.where('LOWER(department.name) LIKE :filter', { filter: `%${filter.toLowerCase()}%` });
+      }
+
+      queryBuilder.orderBy(`department.${sortBy}`, sortOrder).skip(skip).take(limit);
+
+      queryBuilder.leftJoinAndSelect('department.unit', 'unit');  // Carregar a relação 'unit'
+
+      const [departments, total] = await queryBuilder.getManyAndCount();
       return { data: departments, total };
     } catch (error) {
-      this.logger.error('Erro ao buscar departamentos', error.stack);
-      throw new Error(error);
+      throw new Error(`Erro ao buscar departamentos com paginação: ${error.message}`);
     }
   }
 
-  async getDepartmentsByUnit(unitId: string, page: number, limit: number): Promise<{ data: DepartmentEntity[], total: number }> {
+  // Método para buscar todos os departamentos sem paginação
+  async getAllDepartmentsWithoutPagination(): Promise<DepartmentEntity[]> {
     try {
-      this.logger.log(`Buscando departamentos pela unidade: ${unitId}`);
-      const [departments, total] = await this.departmentRepository.findAndCount({
-        where: { unit: { id: unitId } },
-        skip: (page - 1) * limit,
-        take: limit,
+      const departments = await this.departmentRepository.find({
         relations: ['unit'],
       });
+      return departments;
+    } catch (error) {
+      throw new Error('Erro ao buscar todos os departamentos');
+    }
+  }
+
+  // Busca de departamentos por unidade
+  async getDepartmentsByUnit(
+    unitId: string
+  ): Promise<{ data: DepartmentEntity[]}> {
+    try {
+      this.logger.log(`Buscando departamentos pela unidade: ${unitId}`);
+      
+      const [departments] = await this.departmentRepository.findAndCount({
+        where: { unit: { id: unitId } }, 
+        relations: ['unit'], 
+      });
+      
       this.logger.log(`Departamentos encontrados: ${departments.length}`);
-      return { data: departments, total };
+      return { data: departments };
     } catch (error) {
       this.logger.error('Erro ao buscar departamentos pela unidade', error.stack);
       throw new Error(error);
     }
   }
 
-
+  // Busca de departamento por ID
   async getDepartmentById(id: string): Promise<DepartmentEntity> {
     try {
       return await this.departmentRepository.findOne({ where: { id }, relations: ['unit'] });
@@ -98,12 +136,20 @@ export class DepartmentsService {
     }
   }
 
+  // Exclusão de departamento
   async deleteDepartment(id: string): Promise<void> {
     try {
+      const department = await this.getDepartmentById(id);
+      if (!department) {
+        throw new NotFoundException('Departamento não encontrado');
+      }
       await this.departmentRepository.delete(id);
     } catch (error) {
       this.logger.error('Erro ao deletar departamento', error.stack);
-      throw new Error(error);
+      if (error instanceof QueryFailedError && error.message.includes('violates foreign key constraint')) {
+        throw new BadRequestException('Não é possível deletar o departamento porque ele está vinculado a outros registros.');
+      }
+      throw new BadRequestException('Erro ao deletar departamento');
     }
   }
 }
